requirements.txt
# Core dependencies
grpcio==1.62.1
grpcio-tools==1.62.1
python-dotenv==1.0.1
transformers==4.38.2
torch==2.2.1
urllib3==1.26.20

# Development dependencies
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 